requirements.txt
flask
plivo
openai
python-dotenv